requirements.txt
daphne>=4.1.0,<4.2
uwsgi>=2.0.24,<2.1.0
psycopg2>=2.9.7,<2.10
Django>=5.0.3,<6.0
django-cors-headers>=4.3.0,<4.4
djangorestframework>=3.15.0,<3.16
channels>=4.0.0,<4.1
redis>=5.0.0,<5.1
channels-redis>=4.2.0,<4.3
drf-spectacular>=0.27.1,<0.28
Pillow>=10.2.0,<10.3.0